Attempt too support both API versions
