+++ pyprusalink/error.py
@@ -1,0 +1,10 @@
+class PrusaLinkError(Exception):
+    """Base class for PrusaLink errors."""
+
+
+class InvalidAuth(PrusaLinkError):
+    """Error to indicate there is invalid auth."""
+
+
+class Conflict(PrusaLinkError):
+    """Error to indicate the command hit a conflict."""
